+++ op-e2e/setup.go
@@ -267,6 +267,15 @@ type System struct {
 	TimeTravelClock *clock.AdvancingClock
 }
 
+// TestAccount defines an account generated by startConfigWithTestAccounts
+type TestAccount struct {
+	HDPath string
+	Key    *ecdsa.PrivateKey
+	Addr   common.Address
+	L1Opts *bind.TransactOpts
+	L2Opts *bind.TransactOpts
+}
+
 func (sys *System) NodeEndpoint(name string) string {
 	return selectEndpoint(sys.EthInstances[name])
 }
