+++ op-e2e/system_tob_test.go
@@ -3,7 +3,6 @@ package op_e2e
 import (
 	"bytes"
 	"context"
-	"crypto/ecdsa"
 	"fmt"
 	"math/big"
 	"math/rand"
@@ -156,15 +155,6 @@ func TestL2SequencerRPCDepositTx(t *testing.T) {
 	err = l2Verif.SendTransaction(ctx, tx)
 	cancel()
 	require.Error(t, err, "a DepositTx was accepted by L2 verifier over RPC when it should not have been.")
-}
-
-// TestAccount defines an account generated by startConfigWithTestAccounts
-type TestAccount struct {
-	HDPath string
-	Key    *ecdsa.PrivateKey
-	Addr   common.Address
-	L1Opts *bind.TransactOpts
-	L2Opts *bind.TransactOpts
 }
 
 // startConfigWithTestAccounts takes a SystemConfig, generates additional accounts, adds them to the config, so they
